my changes

+++ mysite/accounts/views.py
@@ -1,21 +1,46 @@
 
 from django.shortcuts import render, redirect
-from .forms import UserRegistrationForm, VerifyCodeForm
+from .forms import UserRegistrationForm, VerifyCodeForm,LoginForm,UserProfileForm
 from django.views import View
+from django.forms import inlineformset_factory
+from django.views.generic import TemplateView
 import random
 from extensions.utils import send_otp
-from .models import User, OtpCode
+from .models import User, OtpCode,UserProfile
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.utils.decorators import method_decorator
+
+
 # Create your views here.
-class UserRegistrationView(View):
-    form_class = UserRegistrationForm
-    template_name = "accounts/register.html"
+class AuthView(View):
+    login_form_class=LoginForm
+    register_form_class=UserRegistrationForm
+    verify_form_class=VerifyCodeForm
+    template_name = "accounts/login.html"
 
     def get(self, request):
-        form = self.form_class
-        return render(request, self.template_name, {"form": form})
-
+        if "user_registration_info" in request.session:
+            form=self.verify_form_class
+            
+            status="verify"
+            return render(request, self.template_name, {"form": form,"status":status,"registerform":self.register_form_class})
+        else:
+            status="login"
+            form=self.login_form_class
+            return render(request, self.template_name, {"form": form,"status":status,"registerform":self.register_form_class})
+   
     def post(self, request):
-        form = self.form_class(request.POST)
+        if 'register' in request.POST:
+            return self.handle_registration(request)
+        elif 'login' in request.POST:
+            return self.handle_login(request)
+        elif 'verify' in request.POST and "user_registration_info" in request.session:
+            return self.handle_verification(request)
+        return self.get(request)
+        
+    def handle_registration(self, request):
+        form = self.register_form_class(request.POST)
         if form.is_valid():
             code = random.randint(1000, 9999)
             cd = form.cleaned_data
@@ -24,38 +49,70 @@ class UserRegistrationView(View):
             OtpCode.objects.create(phone_number=phone_number, code=code)
             request.session["user_registration_info"] = {
                 "phone_number": phone_number,
-                "email": cd["email"],
-                "first_name": cd["first_name"],
-                "last_name": cd["last_name"],
-                "password": cd["password2"],
-            }
-            return redirect("accounts:verify")
-        return render(request, self.template_name, {"form": form})
-
-
-class UserVerifyView(View):
-    form_class = VerifyCodeForm
-    template_name = "accounts/verify.html"
-
-    def get(self, request):
-        form = self.form_class
-        return render(request, self.template_name, {"form": form})
-
-    def post(self, request):
-        user_session = request.session["user_registration_info"]
-        code_instance = OtpCode.objects.get(phone_number=user_session["phone_number"])
-        form = self.form_class(request.POST)
+                "password": cd["password2"],}
+            return redirect('accounts:login')
+        
+        return render(request, self.template_name, {"registerform": form,"form":form,"status":'register',})
+    
+    def handle_login(self, request):
+        form = self.login_form_class(request.POST)
         if form.is_valid():
-            code = form.cleaned_data["code"]
-            if code == code_instance.code:
-                User.objects.create(
-                    phone_number=user_session["phone_number"],
-                    email=user_session["email"],
-                    first_name=user_session["first_name"],
-                    last_name=user_session["last_name"],
-                    password=user_session["password"],
-                )
-                code_instance.delete()
-                del request.session["user_registration_info"]
+            phone_number = form.cleaned_data["phone_number"]
+            password = form.cleaned_data["password"]
+            user = authenticate(request, phone_number=phone_number, password=password)
+            if user is not None:
+                login(request, user)
                 return redirect("blog:post_list")
-        return render(request, self.template_name, {"form": form})
+        return render(request, self.template_name, {"form": form,"registerform": self.register_form_class,"status":'login'})
+    
+    def handle_verification(self,request):
+            user_session = request.session["user_registration_info"]
+            code_instance = OtpCode.objects.get(phone_number=user_session["phone_number"])
+            if not code_instance:
+                    return render(request,self.template_name,{"error":"invalid session."})
+            form = self.verify_form_class(request.POST)
+            if form.is_valid():
+                code = form.cleaned_data["code"]
+                if code == code_instance.code:
+                    
+                    user=User(phone_number=user_session["phone_number"],)
+                    user.set_password(user_session["password"])
+                    user.save()
+                    login(request, user)
+                    code_instance.delete()
+                    del request.session["user_registration_info"]
+                    return redirect("blog:post_list")
+            return render(request, self.template_name, {"form": form,"registerform": self.register_form_class,"status":'verify'})
+
+
+@login_required
+def profile(request):
+    if request.method == 'POST':
+        
+        profile_form = UserProfileForm(request.POST, instance=request.user.userprofile)
+
+        if profile_form.is_valid():
+            cd = profile_form.cleaned_data
+            userprofile=request.user.userprofile
+            userprofile.status=cd['status']
+            userprofile.companytype=cd['companytype']
+            userprofile.first_name=cd['first_name']
+            userprofile.last_name=cd['last_name']
+            userprofile.company_name=cd['company_name']
+            userprofile.city=cd['city']
+            userprofile.social_media=cd['social_media']
+            userprofile.save()
+         #hame fielda byd poor she
+    else:
+        
+        profile_form = UserProfileForm(instance=request.user.userprofile)
+
+    return render(request, 'accounts/profile.html', {'profile_form': profile_form})
+
+
+class LogoutView(View):
+    def get(self, request):
+        logout(request)
+        return redirect("blog:post_list")
+    
+
